+++ src/shared/ChangeLog.lua
@@ -1,4 +1,12 @@
 return {
+	"2/16/2024",
+	{
+		"Fixed",
+		{
+			"The barrier into the mine now works correctly and allows you through when you defeat the noble.",
+			"Various visual issues throughout the map have been corrected.",
+		},
+	},
 	"2/13/2024 Part 2",
 	{
 		"Added",